git-svn-id: svn://182.53.32.18/bbip/project/GDEUPSB/TRUNK/GDEUPSB@12925 2ac714ae-9a16-11e4-ad83-2f06aba1930a

+++ src/com/bocom/bbip/gdeupsb/strategy/efek/cancelFeeOnline/PreCclToTHDAction.java
@@ -33,7 +33,7 @@ public class PreCclToTHDAction implements Executable {
 			context.setData(ParamKeys.TXN_TME, DateUtils.parse(DateUtils.formatAsTranstime(new Date())));
 			context.setData(ParamKeys.THD_CUS_NO, context.getData(GDParamKeys.PAY_NO));
 			
-			double i=Integer.parseInt(context.getData(ParamKeys.TXN_AMT).toString());
+			double i=Double.parseDouble(context.getData(ParamKeys.TXN_AMT).toString());
 			double d=i/100;
 			DecimalFormat df=new DecimalFormat("#.00");
 			BigDecimal txnAmt=new BigDecimal(df.format(d));
